restructured and added GameCore
I restructured the file tree for organization. And I added GameCore which will contain all core aspects for the game and will prepare to initialize the needed resources. As well as it will have debug usage as well, as the GameCore constructor will take a bool to toggle debug output.